requirements.txt
# 修仙模拟器 - 依赖包列表

# 核心依赖
PyYAML>=6.0  # 配置文件解析

# GUI 模式依赖
customtkinter>=5.2.0  # 现代化 GUI 框架
Pillow>=10.0.0  # 图像处理
pygame>=2.5.0  # 音效支持

# Web 模式依赖
Flask>=3.0.0  # Web 框架
Flask-CORS>=4.0.0  # 跨域支持

# 开发依赖
pytest>=7.4.0  # 测试框架
pytest-cov>=4.1.0  # 测试覆盖率
black>=23.0.0  # 代码格式化
flake8>=6.1.0  # 代码检查
mypy>=1.5.0  # 类型检查
isort>=5.12.0  # 导入排序

# 可选依赖
# rich>=13.0.0  # 增强终端输出（文本模式）
